taiwan-stock-chip-analysis/requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
python-dotenv==1.0.0
sqlite3
schedule==1.2.0
colorama==0.4.6
tqdm==4.65.0